audio_bk_mkr/requirements.txt
streamlit
python-dotenv
pandas
requests
openai
playsound
pydub
